moving constants to specific packages
