+++ src/main/kotlin/com/lmo/ninie/io/constants/commands/CommandManuals.kt
@@ -1,4 +1,4 @@
-package com.lmo.ninie.io.constants
+package com.lmo.ninie.io.constants.commands
 
 object CommandManuals {
     val HELP = """
