+++ src/main/kotlin/com/lmo/ninie/io/constants/text/Emojis.kt
@@ -1,4 +1,4 @@
-package com.lmo.ninie.io.constants
+package com.lmo.ninie.io.constants.text
 
 object Emojis {
     const val HEART = ":heart:"
