+++ src/main/kotlin/com/lmo/ninie/io/extensions/eventmessage/CallsNinie.kt
@@ -1,6 +1,6 @@
 package com.lmo.ninie.io.extensions.eventmessage
 
-import com.lmo.ninie.io.constants.MagicStrings
+import com.lmo.ninie.io.constants.text.MagicStrings
 import discord4j.core.`object`.entity.Message
 
 fun Message.callsNinie(prefix:String) :Boolean{
